refactor: move star wars to own pkg
